# docs/demo3/openapi.yaml
openapi: 3.0.4
info:
  title: FlashID API
  description: South African Digital ID Wallet backend API.
  version: v1
paths:
...
  /api/institutions/register:
    post:
      tags:
      - Institutions
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterInstitutionRequestDto'
          text/json:
            schema:
              $ref: '#/components/schemas/RegisterInstitutionRequestDto'
          application/*+json:
            schema:
              $ref: '#/components/schemas/RegisterInstitutionRequestDto'
      responses:
        '200':
          description: OK
...
  /api/institutions/{institutionId}:
    get:
      tags:
      - Institutions
      parameters:
      - name: institutionId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
...
components:
  schemas:
...
    InstitutionType:
      enum:
      - HomeAffairs
      - LicensingDepartment
      - LawEnforcement
      - Healthcare
      - FinancialInstitution
      type: string
...
    RegisterInstitutionRequestDto:
      type: object
      properties:
        name:
          type: string
          nullable: true
        type:
          $ref: '#/components/schemas/InstitutionType'
        verificationNumber:
          type: string
          nullable: true
        adminId:
          type: string
          format: uuid
      additionalProperties: false
...